Try to fix a use-after-free issue in thread_exit
thread_exit performs a delayed free of the thread structure and its
stack before calling thread_resched. This creates a race condition since
thread_resched still uses some variables in the old thread structure.
Especially critical are the arch_context_switch functions as those might
push new data into the oldthread's thread structure that is already in
the delayed free list. A second core may come in and free this structure
and subsequentially reuse (that is actually how I found it in the first
place).

By moving the free to later in the process, i.e. after the last
interaction with the oldthread solves this issue.